new_tomatoclock/requirements.txt
# 核心依赖
python>=3.8
pygame>=2.1.3
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.22.0

# 数据库
sqlite3>=3.35.0

# 其他工具
python-dateutil>=2.8.2
typing-extensions>=4.0.0

# 开发依赖
black>=22.1.0
flake8>=4.0.1
mypy>=0.930
pytest>=7.0.0
pytest-cov>=3.0.0
